requirements (1).txt
python-telegram-bot==21.0
anthropic>=0.18.0
google-generativeai>=0.3.0
beautifulsoup4>=4.12.0
requests>=2.31.0
python-dotenv>=1.0.0
lxml>=4.9.0
